feature: remove unused files and adjust some design issue (#1)
