--- src/components/footer.tsx
@@ -1,9 +1,0 @@
-import ExtLink from './ext-link'
-
-export default () => (
-  <>
-    <footer>
-      <ExtLink href="https://github.com/ijjk/notion-blog">Source code</ExtLink>
-    </footer>
-  </>
-)
