+++ src/pages/[slug].tsx
@@ -1,15 +1,19 @@
 import React from 'react'
 import Head from 'next/head'
-import Header from '../../components/header'
-import Heading from '../../components/heading'
-import components from '../../components/dynamic'
+import Link from 'next/link'
+import Header from '../components/header'
+import Heading from '../components/heading'
+import components from '../components/dynamic'
 import ReactJSXParser from '@zeit/react-jsx-parser'
-import blogStyles from '../../styles/blog.module.css'
-import { textBlock } from '../../lib/notion/renderers'
-import getPageData from '../../lib/notion/getPageData'
-import getBlogIndex from '../../lib/notion/getBlogIndex'
-import getNotionUsers from '../../lib/notion/getNotionUsers'
-import { getBlogLink, getDateStr } from '../../lib/blog-helpers'
+import { textBlock } from '../lib/notion/renderers'
+import getPageData from '../lib/notion/getPageData'
+import getBlogIndex from '../lib/notion/getBlogIndex'
+import getNotionUsers from '../lib/notion/getNotionUsers'
+import { getBlogLink, getDateStr } from '../lib/blog-helpers'
+
+// styles
+import sharedStyles from '../styles/shared.module.css'
+import blogStyles from '../styles/blog.module.css'
 
 // Get the data for each blog post
 export async function unstable_getStaticProps({ params: { slug } }) {
@@ -66,15 +70,15 @@ const RenderPost = ({ post, redirect }) => {
 
   return (
     <>
-      <Header titlePre={post.Page} />
+      <Header />
+      <div className={`${sharedStyles.layout} ${sharedStyles.textLink}`}>
+        <h1>
+          <Link href="/">omuomugin-pub</Link>
+        </h1>
+      </div>
       <div className={blogStyles.post}>
         <h1>{post.Page || ''}</h1>
-        {post.Authors.length > 0 && (
-          <div className="authors">By: {post.Authors.join(' ')}</div>
-        )}
-        {post.Date && (
-          <div className="posted">Posted: {getDateStr(post.Date)}</div>
-        )}
+        {post.Date && <div className="posted">{getDateStr(post.Date)}</div>}
 
         <hr />
