+++ src/pages/_app.tsx
@@ -1,26 +1,51 @@
 import '../styles/global.css'
+import contactStyles from '../styles/contact.module.css'
+
 import ExtLink from '../components/ext-link'
+
+import GitHub from '../components/svgs/github'
+import Twitter from '../components/svgs/twitter'
+import Envelope from '../components/svgs/envelope'
+import LinkedIn from '../components/svgs/linkedin'
+
+const contacts = [
+  {
+    Comp: Twitter,
+    alt: 'twitter icon',
+    link: 'https://twitter.com/omuomugin',
+  },
+  {
+    Comp: GitHub,
+    alt: 'github icon',
+    link: 'https://github.com/omuomugin',
+  },
+  {
+    Comp: Envelope,
+    alt: 'envelope icon',
+    link: 'omuomugin@out-box.dev',
+  },
+]
 
 export default ({ Component, pageProps }) => (
   <>
     <Component {...pageProps} />
 
     <footer>
-      <span>Deploy your own!</span>
-      <ExtLink href="https://zeit.co/new/project?template=https://github.com/ijjk/notion-blog/tree/master">
-        <img
-          src="https://zeit.co/button"
-          height={46}
-          width={132}
-          alt="deploy to ZEIT button"
-        />
-      </ExtLink>
-      <span>
-        or{' '}
-        <ExtLink href="https://github.com/ijjk/notion-blog">
-          view source
+      <div className={contactStyles.links}>
+        {contacts.map(({ Comp, link, alt }) => {
+          return (
+            <ExtLink key={link} href={link} aria-label={alt}>
+              <Comp height={32} />
+            </ExtLink>
+          )
+        })}
+      </div>
+      <div>
+        Hosted using{' '}
+        <ExtLink href="https://github.com/omuomugin/notion-blog">
+          notion-blog
         </ExtLink>
-      </span>
+      </div>
     </footer>
   </>
 )
