+++ src/styles/global.css
@@ -198,40 +198,6 @@ pre.dark[class*='language-'] {
   align-items: center;
 }
 
-.features {
-  display: flex;
-  flex-wrap: wrap;
-  margin: 4rem auto;
-  width: 920px;
-  max-width: calc(100vw - var(--gap-double));
-}
-
-.feature {
-  flex: 0 0 25%;
-  align-items: center;
-  display: inline-flex;
-  padding: 0 0.5rem 1.5rem;
-  margin: 0 auto;
-}
-
-.feature h4 {
-  margin: 0 0 0 0.5rem;
-  font-weight: 700;
-  font-size: 0.95rem;
-  white-space: nowrap;
-}
-
-@media (max-width: 600px) {
-  .feature div {
-    flex-basis: auto;
-    padding-left: 0;
-  }
-
-  .feature h4 {
-    font-size: 0.75rem;
-  }
-}
-
 .explanation {
   font-size: 1rem;
   width: 35rem;
@@ -279,7 +245,7 @@ mark {
 
 h1 {
   margin: var(--gap-double) 0 calc(0.5 * var(--gap)) 0;
-  font-size: 2.25rem;
+  font-size: 1.5rem;
   font-weight: 800;
   letter-spacing: -0.05rem;
 }
